requirements.txt
# USB Microscope Driver Dependencies

# USB Communication
pyusb>=1.2.1

# Image Processing
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0

# GUI (Optional)
# tkinter is included in Python standard library

# Development Tools (Optional)
matplotlib>=3.5.0  # For data visualization